Adjust json_manifest_per_file_callback API in one more place.
Oversight in commit d94cf5ca7 (and in my testing of same).

Discussion: https://postgr.es/m/9468.1727895630@sss.pgh.pa.us

+++ src/backend/backup/basebackup_incremental.c
@@ -58,7 +58,7 @@ typedef struct
 {
 	uint32		status;
 	const char *path;
-	size_t		size;
+	uint64		size;
 } backup_file_entry;
 
 static uint32 hash_string_pointer(const char *s);
@@ -133,7 +133,7 @@ static void manifest_process_system_identifier(JsonManifestParseContext *context
 											   uint64 manifest_system_identifier);
 static void manifest_process_file(JsonManifestParseContext *context,
 								  const char *pathname,
-								  size_t size,
+								  uint64 size,
 								  pg_checksum_type checksum_type,
 								  int checksum_length,
 								  uint8 *checksum_payload);
@@ -966,7 +966,7 @@ manifest_process_system_identifier(JsonManifestParseContext *context,
  */
 static void
 manifest_process_file(JsonManifestParseContext *context,
-					  const char *pathname, size_t size,
+					  const char *pathname, uint64 size,
 					  pg_checksum_type checksum_type,
 					  int checksum_length,
 					  uint8 *checksum_payload)
